+++ test/apple/code-tools.test.ts
@@ -429,7 +429,7 @@ describe('code-tools', () => {
     });
   });
 
-  describe('#findAppDidFinishLaunchingWithOptions', () => {
+  describe('#findAppDidFinishLaunchingWithOptionsInDirectory', () => {
     describe('no files given', () => {
       it('should check files in directory', () => {
         // -- Arrange --
@@ -442,7 +442,9 @@ describe('code-tools', () => {
 
         // -- Act --
         const result =
-          exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+          exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+            tempDir,
+          );
 
         // -- Assert --
         expect(result).toBe(filePath);
@@ -462,7 +464,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBe(filePath);
@@ -477,7 +481,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBeNull();
@@ -498,7 +504,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBe(filePath);
@@ -513,7 +521,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBeNull();
@@ -534,7 +544,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBe(filePath);
@@ -549,7 +561,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBeNull();
@@ -569,10 +583,9 @@ describe('code-tools', () => {
           );
 
           // -- Act --
-          const result = exportForTesting.findAppDidFinishLaunchingWithOptions(
-            tempDir,
-            [filePath],
-          );
+          const result = exportForTesting.findAppDidFinishLaunchingWithOptions([
+            filePath,
+          ]);
 
           // -- Assert --
           expect(result).toBe(filePath);
@@ -587,7 +600,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBeNull();
@@ -606,10 +621,9 @@ describe('code-tools', () => {
         );
 
         // -- Act --
-        const result = exportForTesting.findAppDidFinishLaunchingWithOptions(
-          tempDir,
-          [filePath],
-        );
+        const result = exportForTesting.findAppDidFinishLaunchingWithOptions([
+          filePath,
+        ]);
 
         // -- Assert --
         expect(result).toBeNull();
@@ -627,10 +641,9 @@ describe('code-tools', () => {
         );
 
         // -- Act --
-        const result = exportForTesting.findAppDidFinishLaunchingWithOptions(
-          tempDir,
-          [filePath],
-        );
+        const result = exportForTesting.findAppDidFinishLaunchingWithOptions([
+          filePath,
+        ]);
 
         // -- Assert --
         expect(result).toBeNull();
@@ -650,7 +663,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBeNull();
@@ -668,7 +683,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBeNull();
@@ -686,7 +703,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBeNull();
@@ -702,7 +721,9 @@ describe('code-tools', () => {
 
           // -- Act --
           const result =
-            exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+            exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+              tempDir,
+            );
 
           // -- Assert --
           expect(result).toBeNull();
@@ -723,7 +744,9 @@ describe('code-tools', () => {
 
         // -- Act --
         const result =
-          exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+          exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+            tempDir,
+          );
 
         // -- Assert --
         expect(result).toBe(filePath);
@@ -749,7 +772,9 @@ describe('code-tools', () => {
 
         // -- Act --
         const result =
-          exportForTesting.findAppDidFinishLaunchingWithOptions(tempDir);
+          exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+            tempDir,
+          );
 
         // -- Assert --
         expect(result).toBe(nestedFilePath);
@@ -764,10 +789,10 @@ describe('code-tools', () => {
         );
 
         // -- Act --
-        const result = exportForTesting.findAppDidFinishLaunchingWithOptions(
-          tempDir,
-          [],
-        );
+        const result =
+          exportForTesting.findAppDidFinishLaunchingWithOptionsInDirectory(
+            tempDir,
+          );
 
         // -- Assert --
         expect(result).toBeNull();
@@ -778,12 +803,8 @@ describe('code-tools', () => {
   describe('#addCodeSnippetToProject', () => {
     describe('app delegate file is not found', () => {
       it('should return false', () => {
-        // -- Arrange --
-        const tempDir = prepareTempDir();
-
         // -- Act --
         const result = addCodeSnippetToProject(
-          tempDir,
           ['AppDelegate.swift'],
           'https://example.com/sentry-dsn',
         );
@@ -820,7 +841,7 @@ describe('code-tools', () => {
 
           it('should add the code snippet', () => {
             // -- Act --
-            const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+            const result = addCodeSnippetToProject([filePath], dsn);
 
             // -- Assert --
             expect(result).toBeTruthy();
@@ -830,7 +851,7 @@ describe('code-tools', () => {
 
           it("should set tag 'code-language'", () => {
             // -- Act --
-            const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+            const result = addCodeSnippetToProject([filePath], dsn);
 
             // -- Assert --
             expect(result).toBeTruthy();
@@ -842,7 +863,7 @@ describe('code-tools', () => {
 
           it("should set tag 'ui-engine'", () => {
             // -- Act --
-            const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+            const result = addCodeSnippetToProject([filePath], dsn);
 
             // -- Assert --
             expect(result).toBeTruthy();
@@ -861,7 +882,7 @@ describe('code-tools', () => {
             );
 
             // -- Act --
-            const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+            const result = addCodeSnippetToProject([filePath], dsn);
 
             // -- Assert --
             expect(result).toBeTruthy();
@@ -886,7 +907,7 @@ describe('code-tools', () => {
 
             it('should add the code snippet', () => {
               // -- Act --
-              const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+              const result = addCodeSnippetToProject([filePath], dsn);
 
               // -- Assert --
               expect(result).toBeTruthy();
@@ -898,7 +919,7 @@ describe('code-tools', () => {
 
             it("should set tag 'code-language'", () => {
               // -- Act --
-              const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+              const result = addCodeSnippetToProject([filePath], dsn);
 
               // -- Assert --
               expect(result).toBeTruthy();
@@ -911,7 +932,7 @@ describe('code-tools', () => {
 
             it("should set tag 'ui-engine'", () => {
               // -- Act --
-              const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+              const result = addCodeSnippetToProject([filePath], dsn);
 
               // -- Assert --
               expect(result).toBeTruthy();
@@ -934,7 +955,7 @@ describe('code-tools', () => {
               );
 
               // -- Act --
-              const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+              const result = addCodeSnippetToProject([filePath], dsn);
 
               // -- Assert --
               expect(result).toBeTruthy();
@@ -957,7 +978,7 @@ describe('code-tools', () => {
             );
 
             // -- Act --
-            const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+            const result = addCodeSnippetToProject([filePath], dsn);
 
             // -- Assert --
             expect(result).toBeFalsy();
@@ -977,7 +998,7 @@ describe('code-tools', () => {
             );
 
             // -- Act --
-            const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+            const result = addCodeSnippetToProject([filePath], dsn);
 
             // -- Assert --
             expect(result).toBeTruthy();
@@ -1001,7 +1022,7 @@ describe('code-tools', () => {
 
           it('should not add the code snippet', () => {
             // -- Act --
-            const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+            const result = addCodeSnippetToProject([filePath], dsn);
 
             // -- Assert --
             expect(result).toBeTruthy();
@@ -1011,7 +1032,7 @@ describe('code-tools', () => {
 
           it('should log info', () => {
             // -- Act --
-            const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+            const result = addCodeSnippetToProject([filePath], dsn);
 
             // -- Assert --
             expect(result).toBeTruthy();
@@ -1031,7 +1052,7 @@ describe('code-tools', () => {
           );
 
           // -- Act --
-          const result = addCodeSnippetToProject(tempDir, [filePath], dsn);
+          const result = addCodeSnippetToProject([filePath], dsn);
 
           // -- Assert --
           expect(result).toBeTruthy();